hc-api/requirements.txt
uuid
configparser
functools
pyjwt
flask
flask_marshmallow
flask_sqlalchemy